+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 > - Breaking Changes:
 > - Features:
 > - Bugfixes:
+>	- Fixed an issue where connections would not gain focus when selected, which could prevent editor keybindings from functioning in certain scenarios
 >	- Resolved an issue where selecting a node did not deselect connections and vice versa
 
 #### **Version 7.0.0**