+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 > - Bugfixes:
 >	- Fixed an issue where connections would not gain focus when selected, which could prevent editor keybindings from functioning in certain scenarios
 >	- Resolved an issue where selecting a node did not deselect connections and vice versa
+>	- Fixed a bug preventing ItemContainers from being dragged when the mouse could not be captured
 
 #### **Version 7.0.0**